[luci] Extract QuantizeWeights from QuantizeWithMinMax
This extracts QuantizeWeights from QuantizeWithMinMax.

ONE-DCO-1.0-Signed-off-by: Hyukjin Jeong <hj1.jeong@samsung.com>
